🔧 refactor: 메시지 템플릿 구조 개선 및 불필요한 파일 제거
