+++ services/api/app/src/api/public/cafeterias/cafeterias.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { CafeteriaMessagesService } from 'src/api/public/cafeterias/cafeteria-messages.service';
 import { CampusesModule } from 'src/api/public/campuses/campuses.module';
-import { CampusMessagesService } from 'src/api/public/message-templates/campus-messages.service';
+import { CampusMessagesService } from 'src/api/public/campuses/campus-messages.service';
 import { CafeteriasRepositoryModule } from 'src/type-orm/entities/cafeterias/cafeterias-repository.module';
 import { CafeteriasController } from './cafeterias.controller';
 import { CafeteriasService } from './cafeterias.service';
