+++ services/api/app/src/api/public/notices/notices.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { CommonMessagesService } from 'src/api/public/message-templates/common-messages.service';
-import { NoticeMessagesService } from 'src/api/public/message-templates/notice-messages.service';
+import { CommonMessagesService } from 'src/api/public/common/common-messages.service';
+import { NoticeMessagesService } from 'src/api/public/notices/notice-messages.service';
 import { NoticesRepositoryModule } from 'src/type-orm/entities/notices/notices-repository.module';
 import { NoticesController } from './notices.controller';
 import { NoticesService } from './notices.service';
