+++ services/api/app/src/api/public/schedules/schedules.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { CommonMessagesService } from 'src/api/public/message-templates/common-messages.service';
-import { ScheduleMessagesService } from 'src/api/public/message-templates/schedule-messages.service';
+import { CommonMessagesService } from 'src/api/public/common/common-messages.service';
+import { ScheduleMessagesService } from 'src/api/public/schedules/schedule-messages.service';
 import { AcademicCalendarsRepositoryModule } from 'src/type-orm/entities/academic-calendars/academic-calendars-repository.module';
 import { SchedulesController } from 'src/api/public/schedules/schedules.controller';
 import { SchedulesService } from './schedules.service';
