+++ services/api/app/src/api/public/users/users.controller.ts
@@ -8,11 +8,11 @@ import { BlockId } from 'src/api/common/utils/constants';
 import { CampusesService } from 'src/api/public/campuses/campuses.service';
 import { CollegesService } from 'src/api/public/colleges/colleges.service';
 import { DepartmentsService } from 'src/api/public/departments/departments.service';
-import { CampusMessagesService } from 'src/api/public/message-templates/campus-messages.service';
-import { CollegeMessagesService } from 'src/api/public/message-templates/college-messages.service';
-import { CommonMessagesService } from 'src/api/public/message-templates/common-messages.service';
-import { DepartmentMessagesService } from 'src/api/public/message-templates/department-messages.service';
-import { UserMessageService } from 'src/api/public/message-templates/user-messages.service';
+import { CampusMessagesService } from 'src/api/public/campuses/campus-messages.service';
+import { CollegeMessagesService } from 'src/api/public/colleges/college-messages.service';
+import { CommonMessagesService } from 'src/api/public/common/common-messages.service';
+import { DepartmentMessagesService } from 'src/api/public/departments/department-messages.service';
+import { UserMessageService } from 'src/api/public/users/user-messages.service';
 import { User } from '../../../type-orm/entities/users/users.entity';
 import { CurrentUser } from './decorators/current-user.decorator';
 import { FetchCurrentUser } from './decorators/fetch-current-user.decorator';
